requirements.txt

PennyLane
qiskit
torch
torchvision
Pillow
astropy
